[TM-1319] Add HotJar and GoogleAnalytics to the Dashboard
